+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Added `--force` flag to `Tester` for adding or updating testcases to ignore errors if set. (@ddash-ct)
+
 ### Fixed
 - `pefileutils.obtain_export_list` would contain a `null` entry as the last item in the list for any file